move get config to actions
